# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Community Quiz API",
        "description": "Community-Based Quiz Management System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/auth/{provider}/redirect": {
            "get": {
                "tags": [
                    "Social Auth"
                ],
                "summary": "Redirect to OAuth provider",
                "description": "Initiates the OAuth flow for the specified provider.",
                "operationId": "d6c0d543c872b85501bdf7471e725df9",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "OAuth provider name (google, github)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "github"
                            ]
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to provider"
                    }
                }
            }
        },
        "/api/auth/{provider}/callback": {
            "get": {
                "tags": [
                    "Social Auth"
                ],
                "summary": "Handle OAuth provider callback",
                "operationId": "7d9d4331d1865a5482333f5aded233e1",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "OAuth provider name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "github"
                            ]
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to frontend with token"
                    }
                }
            }
        },
...
        "/api/dashboard/quiz-maker": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get statistics for Quiz Maker dashboard",
                "operationId": "dashboardQuizMaker",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_categories": {
                                            "type": "integer"
                                        },
                                        "total_quizzes": {
                                            "type": "integer"
                                        },
                                        "total_communities": {
                                            "type": "integer"
                                        },
                                        "total_submissions": {
                                            "type": "integer"
                                        },
                                        "pending_reviews": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}